.context/agentfs-main/sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentfs-sdk"
version = "0.6.0"
description = "AgentFS Python SDK - A filesystem and key-value store for AI agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["ai", "agent", "turso", "sqlite", "key-value", "filesystem"]
authors = [
    {name = "Turso"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    'Operating System :: POSIX :: Linux',
    'Operating System :: Microsoft :: Windows',
    'Operating System :: MacOS',
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyturso==0.4.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.0",
    "ty>=0.0.1a34",
]

[project.urls]
Homepage = "https://github.com/tursodatabase/agentfs"
Source = "https://github.com/tursodatabase/agentfs"
Repository = "https://github.com/tursodatabase/agentfs"
Issues = "https://github.com/tursodatabase/agentfs/issues"

[tool.setuptools.packages.find]
include = ["agentfs_sdk*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
